requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
gunicorn==20.0.4
Jinja2==2.11.1
MarkupSafe==1.1.1
progressbar2==3.47.0
psycopg2==2.8.4
python-decouple==3.3
python-utils==2.3.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
